MultiDBAPI/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
blacksheep==1.2.18
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
essentials==1.1.5
essentials-openapi==1.0.8
filelock==3.12.2
frozenlist==1.4.0
future==0.18.3
geographiclib==2.0
geopy==2.3.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
gspread==5.10.0
guardpost==0.0.9
h11==0.14.0
httptools==0.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
parsel==1.8.1
Protego==0.2.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydig==0.4.0
PyDispatcher==2.0.7
PyJWT==2.8.0
pymailcheck==1.1.0
pyOpenSSL==23.2.0
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-whois==0.8.0
pytz==2023.3
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rodi==1.1.3
rsa==4.9
Scrapy==2.9.0
service-identity==23.1.0
six==1.16.0
SQLAlchemy==2.0.19
street-address==0.4.0
tldextract==3.4.4
twilio==8.5.0
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
w3lib==2.1.1
yarl==1.9.2
zope.interface==6.0
